Corrected All Dao Layers

+++ src/main/java/com/townscript/forum/model/user/UserHibernate.java
@@ -8,7 +8,7 @@ public class UserHibernate {
 	
 	@Id @GeneratedValue
 	@Column(name="USER_ID")
-	private Integer userId;
+	private Long userId;
 	
 	@Column(name="USER_NAME")
 	private String userName;
@@ -31,7 +31,7 @@ public class UserHibernate {
 	public UserHibernate() {
 	}
 
-	public UserHibernate(Integer userId, String userName, String password, String userEmail, String userMobile,
+	public UserHibernate(Long userId, String userName, String password, String userEmail, String userMobile,
 			String userDateTime, byte[] userPropic) {
 		super();
 		this.userId = userId;
@@ -43,11 +43,11 @@ public class UserHibernate {
 		this.userPropic = userPropic;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
